PoC for metics starts
