+++ catalyst/core/callbacks/logging.py
@@ -1,19 +1,52 @@
-from typing import Dict, List  # isort:skip
+from typing import Any, Dict, List  # isort:skip
+from collections import defaultdict
 import logging
 import os
 import sys
-from urllib.parse import quote_plus
-from urllib.request import Request, urlopen
 
 from tqdm import tqdm
 
 from catalyst import utils
-from catalyst.core import _State, LoggerCallback
+from catalyst.core import _State, Callback, CallbackOrder, CallbackNode
 from catalyst.utils.tools.tensorboard import SummaryWriter
 from . import formatters
 
 
-class VerboseLogger(LoggerCallback):
+class MetricManagerCallback(Callback):
+    """
+    Prepares the metrics for the logging, transferring from PyTorch to numpy
+    """
+    def __init__(self):
+        super().__init__(
+            order=CallbackOrder.Logging - 1,
+            node=CallbackNode.All
+        )
+
+    @staticmethod
+    def _to_single_value(value: Any) -> float:
+        if hasattr(value, "item"):
+            value = value.item()
+
+        value = float(value)
+        return value
+
+    @staticmethod
+    def _process_metrics(metrics: Dict[str, Any]):
+        output = {}
+        for key, value in metrics.items():
+            value = MetricManagerCallback._to_single_value(value)
+            value = utils.distributed_mean(value)
+            output[key] = value
+        return output
+
+    def on_batch_start(self, state: _State):
+        state.batch_metrics = defaultdict(None)
+
+    def on_batch_end(self, state: _State):
+        state.batch_metrics = self._process_metrics(state.batch_metrics)
+
+
+class VerboseLogger(Callback):
     """
     Logs the params into console
     """
@@ -29,7 +62,7 @@ class VerboseLogger(LoggerCallback):
                 to remove always_show metrics set it to an empty list ``[]``
             never_show (List[str]): list of metrics which will not be shown
         """
-        super().__init__()
+        super().__init__(order=CallbackOrder.Logging, node=CallbackNode.Master)
         self.tqdm: tqdm = None
         self.step = 0
         self.always_show = (
@@ -72,7 +105,7 @@ class VerboseLogger(LoggerCallback):
         self.tqdm.set_postfix(
             **{
                 k: "{:3.3f}".format(v) if v > 1e-3 else "{:1.3e}".format(v)
-                for k, v in sorted(state.metric_manager.batch_values.items())
+                for k, v in sorted(state.batch_metrics.items())
                 if self._need_show(k)
             }
         )
@@ -95,14 +128,14 @@ class VerboseLogger(LoggerCallback):
             state.need_exception_reraise = False
 
 
-class ConsoleLogger(LoggerCallback):
+class ConsoleLogger(Callback):
     """
     Logger callback,
     translates ``state.metric_manager`` to console and text file
     """
     def __init__(self):
         """Init ``ConsoleLogger``"""
-        super().__init__()
+        super().__init__(order=CallbackOrder.Logging, node=CallbackNode.Master)
         self.logger = None
 
     @staticmethod
@@ -153,7 +186,7 @@ class ConsoleLogger(LoggerCallback):
         self.logger.info("", extra={"state": state})
 
 
-class TensorboardLogger(LoggerCallback):
+class TensorboardLogger(Callback):
     """
     Logger callback, translates ``state.metric_manager`` to tensorboard
     """
@@ -170,7 +203,7 @@ class TensorboardLogger(LoggerCallback):
             log_on_batch_end (bool): logs per-batch metrics if set True
             log_on_epoch_end (bool): logs per-epoch metrics if set True
         """
-        super().__init__()
+        super().__init__(order=CallbackOrder.Logging, node=CallbackNode.Master)
         self.metrics_to_log = metric_names
         self.log_on_batch_end = log_on_batch_end
         self.log_on_epoch_end = log_on_epoch_end
@@ -242,128 +275,11 @@ class TensorboardLogger(LoggerCallback):
             logger.close()
 
 
-class TelegramLogger(LoggerCallback):
-    """
-    Logger callback, translates ``state.metric_manager`` to telegram channel
-    """
-    def __init__(
-        self,
-        token: str = None,
-        chat_id: str = None,
-        metric_names: List[str] = None,
-        log_on_stage_start: bool = True,
-        log_on_loader_start: bool = True,
-        log_on_loader_end: bool = True,
-        log_on_stage_end: bool = True,
-        log_on_exception: bool = True,
-    ):
-        """
-        Args:
-            token (str): telegram bot's token,
-                see https://core.telegram.org/bots
-            chat_id (str): Chat unique identifier
-            metric_names: List of metric names to log.
-                if none - logs everything.
-            log_on_stage_start (bool): send notification on stage start
-            log_on_loader_start (bool): send notification on loader start
-            log_on_loader_end (bool): send notification on loader end
-            log_on_stage_end (bool): send notification on stage end
-            log_on_exception (bool): send notification on exception
-        """
-        super().__init__()
-        # @TODO: replace this logic with global catalyst config at ~/.catalyst
-        self._token = token or os.environ.get("CATALYST_TELEGRAM_TOKEN", None)
-        self._chat_id = (
-            chat_id or os.environ.get("CATALYST_TELEGRAM_CHAT_ID", None)
-        )
-        assert self._token is not None and self._chat_id is not None
-        self._base_url = (
-            f"https://api.telegram.org/bot{self._token}/sendMessage"
-        )
 
-        self.log_on_stage_start = log_on_stage_start
-        self.log_on_loader_start = log_on_loader_start
-        self.log_on_loader_end = log_on_loader_end
-        self.log_on_stage_end = log_on_stage_end
-        self.log_on_exception = log_on_exception
-
-        self.metrics_to_log = metric_names
-
-    def _send_text(self, text: str):
-        try:
-            url = (
-                f"{self._base_url}?"
-                f"chat_id={self._chat_id}&"
-                f"disable_web_page_preview=1&"
-                f"text={quote_plus(text, safe='')}"
-            )
-
-            request = Request(url)
-            urlopen(request)
-        except Exception as e:
-            logging.getLogger(__name__).warning(f"telegram.send.error:{e}")
-
-    def on_stage_start(self, state: _State):
-        """Notify about starting a new stage"""
-        if self.log_on_stage_start:
-            text = f"{state.stage_name} stage was started"
-
-            self._send_text(text)
-
-    def on_loader_start(self, state: _State):
-        """Notify about starting running the new loader"""
-        if self.log_on_loader_start:
-            text = f"{state.loader_name} {state.epoch} epoch was started"
-
-            self._send_text(text)
-
-    def on_loader_end(self, state: _State):
-        """Translate ``state.metric_manager`` to telegram channel"""
-        if self.log_on_loader_end:
-            metrics = state.metric_manager.epoch_values[state.loader_name]
-
-            if self.metrics_to_log is None:
-                metrics_to_log = sorted(list(metrics.keys()))
-            else:
-                metrics_to_log = self.metrics_to_log
-
-            rows: List[str] = [
-                f"{state.loader_name} {state.epoch} epoch was finished:"
-            ]
-
-            for name in metrics_to_log:
-                if name in metrics:
-                    rows.append(utils.format_metric(name, metrics[name]))
-
-            text = "\n".join(rows)
-
-            self._send_text(text)
-
-    def on_stage_end(self, state: _State):
-        """Notify about finishing a stage"""
-        if self.log_on_stage_end:
-            text = f"{state.stage_name} stage was finished"
-
-            self._send_text(text)
-
-    def on_exception(self, state: _State):
-        """Notify about raised Exception"""
-        if self.log_on_exception:
-            exception = state.exception
-            if utils.is_exception(exception) and not isinstance(
-                exception, KeyboardInterrupt
-            ):
-                text = (
-                    f"`{type(exception).__name__}` exception was raised:\n"
-                    f"{exception}"
-                )
-
-                self._send_text(text)
 
 
 __all__ = [
     "ConsoleLogger",
-    "TelegramLogger",
     "TensorboardLogger",
     "VerboseLogger",
 ]
